Refactor into a Jvm Thread
Each Jvm can have multiple threads of execution. Call those
as they are rather than calling them Vms. Also, refactor
the method area and remove the stack. It is not specified
to be part of the structure. Instead of explicitly tracking
a stack of frames during execution of the Jvm thread, use
the system's native stack.